work-at-olist/workatolist/urls.py
from django.conf.urls import url, include
from django.contrib import admin

urlpatterns = [
    url(r'^admin/', admin.site.urls),
    url(r'^api/1/', include('core.urls',
                            namespace='core')),
]
